split routes to own file. fixed linking in dev mode
