+++ docs/javascript/index.html
@@ -4,7 +4,9 @@
   <head>
     <title>My React App</title>
     <meta name="viewport" content="width=device-width,height=device-height,initial-scale=1.0,minimum-scale=1.0">
-    <link rel="stylesheet" href="/node-blog/static/css/main.css"/>
+    
+      <link rel="stylesheet" href="/static/css/main.css"/>
+    
   </head>
   <body>
 
@@ -110,9 +112,9 @@
       </li>
     
       <li>
-        <a href="/node-blog/post/watchMeNodemon">
+        <a href="/node-blog/post/testing-this-jazz">
           <date>March 8, 2017</date>
-          <h2>Nodemon!</h2>
+          <h2>TEsting this ish like whoa!</h2>
           <p>Some fly ass shit!</p>
           <ul>
             
@@ -127,9 +129,9 @@
       </li>
     
       <li>
-        <a href="/node-blog/post/testing-this-jazz">
+        <a href="/node-blog/post/watchMeNodemon">
           <date>March 8, 2017</date>
-          <h2>TEsting this ish like whoa!</h2>
+          <h2>Nodemon!</h2>
           <p>Some fly ass shit!</p>
           <ul>
