+++ docs/post/bootstr-app.html
@@ -4,7 +4,9 @@
   <head>
     <title>My React App</title>
     <meta name="viewport" content="width=device-width,height=device-height,initial-scale=1.0,minimum-scale=1.0">
-    <link rel="stylesheet" href="/node-blog/static/css/main.css"/>
+    
+      <link rel="stylesheet" href="/static/css/main.css"/>
+    
   </head>
   <body>
