+++ src/views/partials/head.ejs
@@ -3,6 +3,10 @@
   <head>
     <title>My React App</title>
     <meta name="viewport" content="width=device-width,height=device-height,initial-scale=1.0,minimum-scale=1.0">
-    <link rel="stylesheet" href="/<%-process.env.npm_package_reponame%>/static/css/main.css"/>
+    <% if (process.env.NODE_ENV === 'production') { %>
+      <link rel="stylesheet" href="/<%-process.env.npm_package_reponame%>/static/css/main.css"/>
+    <% } else{ %>
+      <link rel="stylesheet" href="/static/css/main.css"/>
+    <% } %>
   </head>
   <body>
